:recycle: refactor: 도메인 중심의 패키지 구조로 재구성 (#8)   - place: 장소 관련 도메인 (CommonPlace, PlaceStatistics)   - user: 사용자 및 친구관계 도메인 (FriendRelationship)   - notification : 사용자 알림 관련 도메인   - image: 이미지 처리 관련 도메인   - global: 공통 컴포넌트 (예외처리, API 응답)
