--- src/main/java/com/adit/backend/domain/event/EventRepository.java
@@ -1,8 +1,0 @@
-package com.adit.backend.domain.event;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.adit.backend.domain.event.entity.Event;
-
-public interface EventRepository extends JpaRepository<Event, Long> {
-}
