--- src/main/java/com/adit/backend/domain/image/dto/ImageRequestDto.java
@@ -1,18 +1,0 @@
-package com.adit.backend.domain.image.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ImageRequestDto {
-
-    @NotNull
-    private Long placeId;
-
-    private Long eventId;
-    private String url;
-    private String fileName;
-    private String folderName;
-}
