+++ src/main/java/com/adit/backend/domain/place/entity/UserPlace.java
@@ -1,10 +1,8 @@
-package com.adit.backend.domain.image.entity;
+package com.adit.backend.domain.place.entity;
 
-import com.adit.backend.domain.event.entity.Event;
-import com.adit.backend.domain.place.entity.CommonPlace;
+import com.adit.backend.domain.user.entity.User;
 import com.adit.backend.global.entity.BaseEntity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -20,23 +18,20 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Image extends BaseEntity {
+public class UserPlace extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "place_id")
-    private CommonPlace place;
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "event_id")
-    private Event event;
+    @JoinColumn(name = "common_place_id", nullable = false)
+    private CommonPlace commonPlace;
 
-    @Column(nullable = false)
-    private String url;
-
-    private String fileName;
-    private String folderName;
+    private String memo;
+    private Boolean visited;
 }
