+++ src/main/java/com/adit/backend/domain/place/repository/CommonPlaceRepository.java
@@ -1,0 +1,8 @@
+package com.adit.backend.domain.place.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.adit.backend.domain.place.entity.CommonPlace;
+
+public interface CommonPlaceRepository extends JpaRepository<CommonPlace, Long> {
+}
