+++ src/main/java/com/adit/backend/domain/user/dto/request/FriendRelationshipResponseDto.java
@@ -1,0 +1,23 @@
+package com.adit.backend.domain.user.dto.request;
+
+import com.adit.backend.domain.user.entity.FriendRelationship;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FriendRelationshipResponseDto {
+
+    private Long id;
+    private String status;
+    private Long fromUserId;
+    private Long toUserId;
+
+    public FriendRelationshipResponseDto(FriendRelationship friendRelationship) {
+        this.id = friendRelationship.getId();
+        this.status = friendRelationship.getStatus();
+        this.fromUserId = friendRelationship.getFromUser().getId();
+        this.toUserId = friendRelationship.getToUser().getId();
+    }
+}
