+++ src/main/java/com/adit/backend/domain/user/dto/response/FriendRequestDto.java
@@ -1,0 +1,16 @@
+package com.adit.backend.domain.user.dto.response;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FriendRequestDto {
+
+    @NotNull
+    private Long fromUserId;
+
+    @NotNull
+    private Long toUserId;
+}
