--- src/main/java/com/adit/backend/domain/userPlace/entity/UserPlace.java
@@ -1,38 +1,0 @@
-package com.adit.backend.domain.userPlace.entity;
-
-import com.adit.backend.domain.commonPlace.entity.CommonPlace;
-import com.adit.backend.domain.user.entity.User;
-import com.adit.backend.global.entity.BaseEntity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class UserPlace extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "common_place_id", nullable = false)
-    private CommonPlace commonPlace;
-
-    private String memo;
-    private Boolean visited;
-}
