+++ src/main/java/com/adit/backend/global/error/GlobalExceptionHandler.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.servlet.NoHandlerFoundException;
 
-import com.adit.backend.global.ApiResponse;
+import com.adit.backend.global.common.ApiResponse;
 import com.adit.backend.global.error.exception.GlobalErrorCode;
 
 import lombok.extern.slf4j.Slf4j;
